fix capi, add acpi to doc

+++ capi/modulation_audio_file/audio_file_modulation.h
@@ -3,7 +3,7 @@
 // Created Date: 16/05/2022
 // Author: Shun Suzuki
 // -----
-// Last Modified: 16/05/2022
+// Last Modified: 20/05/2022
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2022 Hapis Lab. All rights reserved.
@@ -16,8 +16,8 @@
 #ifdef __cplusplus
 extern "C" {
 #endif
-EXPORT_AUTD void AUTDModulationRawPCM(void** mod, const char* filename, double sampling_freq, uint16_t mod_sampling_freq_div);
-EXPORT_AUTD void AUTDModulationWav(void** mod, const char* filename, uint16_t mod_sampling_freq_div);
+EXPORT_AUTD void AUTDModulationRawPCM(void** mod, const char* filename, double sampling_freq, uint32_t mod_sampling_freq_div);
+EXPORT_AUTD void AUTDModulationWav(void** mod, const char* filename, uint32_t mod_sampling_freq_div);
 #ifdef __cplusplus
 }
 #endif
