+++ capi/modulation_audio_file/c_api.cpp
@@ -3,7 +3,7 @@
 // Created Date: 16/05/2022
 // Author: Shun Suzuki
 // -----
-// Last Modified: 16/05/2022
+// Last Modified: 20/05/2022
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2022 Hapis Lab. All rights reserved.
@@ -12,12 +12,12 @@
 #include "./audio_file_modulation.h"
 #include "autd3/modulation/audio_file.hpp"
 
-void AUTDModulationRawPCM(void** mod, const char* filename, const double sampling_freq, const uint16_t mod_sampling_freq_div) {
+void AUTDModulationRawPCM(void** mod, const char* filename, const double sampling_freq, const uint32_t mod_sampling_freq_div) {
   const auto filename_ = std::string(filename);
   auto* m = new autd3::modulation::RawPCM(filename_, sampling_freq, mod_sampling_freq_div);
   *mod = m;
 }
-void AUTDModulationWav(void** mod, const char* filename, const uint16_t mod_sampling_freq_div) {
+void AUTDModulationWav(void** mod, const char* filename, const uint32_t mod_sampling_freq_div) {
   const auto filename_ = std::string(filename);
   auto* m = new autd3::modulation::Wav(filename_, mod_sampling_freq_div);
   *mod = m;
